Removed business logic from the user token endpoints

+++ RecipeSocialMediaAPI/Handlers/UserTokens/Notifications/RemoveTokenNotification.cs
@@ -3,6 +3,8 @@ using MongoDB.Bson;
 using RecipeSocialMediaAPI.DAL;
 using RecipeSocialMediaAPI.DAL.Documents;
 using RecipeSocialMediaAPI.DAL.Repositories;
+using RecipeSocialMediaAPI.Endpoints;
+using RecipeSocialMediaAPI.Services;
 using RecipeSocialMediaAPI.Utilities;
 
 namespace RecipeSocialMediaAPI.Handlers.UserTokens.Notifications
@@ -11,15 +13,22 @@ namespace RecipeSocialMediaAPI.Handlers.UserTokens.Notifications
 
     public class RemoveTokenHandler : INotificationHandler<RemoveTokenNotification>
     {
+        private readonly IUserTokenService _userTokenService;
         private readonly IMongoCollectionWrapper<UserTokenDocument> _userTokenCollection;
 
-        public RemoveTokenHandler(IMongoFactory factory, IConfigManager config)
+        public RemoveTokenHandler(IUserTokenService userTokenService, IMongoFactory factory, IConfigManager config)
         {
+            _userTokenService = userTokenService;
             _userTokenCollection = factory.GetCollection<UserTokenDocument>(new UserTokenRepository(), config);
         }
 
         public Task Handle(RemoveTokenNotification notification, CancellationToken cancellationToken)
         {
+            if (!_userTokenService.CheckTokenExistsAndNotExpired(notification.Token))
+            {
+                throw new TokenNotFoundOrExpiredException();
+            }
+
             ObjectId tokenObj = ObjectId.Parse(notification.Token);
             _userTokenCollection.Delete(x => x._id == tokenObj);
